requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-pinecone
sentence-transformers 
langchain-huggingface
pinecone-client
pypdf
opencv-python-headless 
pytesseract 
python-multipart